refactor: move import endpoint from CommonController to CollectionsController
- Move POST /api/import/:name to POST /api/collections/import/:name
- Add FtpService and ImportGatewayService to CollectionsModule
- Remove import endpoint and related imports from CommonController
- Clean up unused imports in CommonController
