+++ src/collections/collections.module.ts
@@ -1,7 +1,9 @@
 import { Module } from '@nestjs/common';
 import { CollectionsController } from './collections.controller';
+import { CommonModule } from '../common/common.module';
 
 @Module({
+  imports: [CommonModule],
   controllers: [CollectionsController],
   providers: [],
 })
