+++ src/common/common.controller.ts
@@ -8,14 +8,10 @@ import {
   UseInterceptors,
   Body,
   UseGuards,
-  Req,
-  Query,
-  BadRequestException,
 } from '@nestjs/common';
 import { MessagePattern, Payload } from '@nestjs/microservices';
-import { ApiBody, ApiConsumes, ApiOperation, ApiParam, ApiQuery, ApiTags } from '@nestjs/swagger';
+import { ApiBody, ApiConsumes, ApiOperation, ApiParam, ApiTags } from '@nestjs/swagger';
 import { FileInterceptor } from '@nestjs/platform-express';
-import multer from 'multer';
 import {
   WhatsappService,
   SmsService,
@@ -24,11 +20,8 @@ import {
   BcbService,
   SmsDto,
   WhatsappDto,
-  ImportGatewayService,
 } from 'src/common';
 import { AuthGuard } from 'src/auth/guards';
-import { ftpStorage } from './services/ftp.service';
-import { Request } from 'express';
 
 @ApiTags('common')
 @Controller('common')
@@ -39,7 +32,6 @@ export class CommonController {
     private readonly whatsapp: WhatsappService,
     private readonly citizenshipDigital: CitizenshipDigitalService,
     private readonly bcbService: BcbService,
-    private readonly importGatewayService: ImportGatewayService,
   ) {}
 
   @MessagePattern('ftp.listFiles')
@@ -88,51 +80,6 @@ export class CommonController {
       message: 'Chunk subido exitosamente',
       serviceStatus: true,
     };
-  }
-
-  @Post('collections/import/:name')
-  @ApiOperation({ summary: 'Importar archivo CSV o Excel' })
-  @ApiConsumes('multipart/form-data')
-  @ApiBody({
-    description: 'Archivo a importar',
-    schema: {
-      type: 'object',
-      properties: {
-        file: {
-          type: 'string',
-          format: 'binary',
-        },
-      },
-    },
-  })
-  @ApiQuery({ name: 'userId', required: false, description: 'Usuario que realiza la importación' })
-  @UseGuards(AuthGuard)
-  async importFile(
-    @Req() req: Request,
-    @Param('name') name: string,
-    @Query('userId') userId?: string,
-  ) {
-    await new Promise<void>((resolve, reject) => {
-      multer({
-        storage: ftpStorage(this.ftp, name),
-        limits: { fileSize: 100 * 1024 * 1024 },
-      }).single('file')(req, {} as any, (err: any) => {
-        if (err) {
-          if (err.code === 'LIMIT_FILE_SIZE') {
-            reject(new BadRequestException('El archivo excede el tamaño máximo permitido (100MB)'));
-          } else {
-            reject(new BadRequestException(err.message));
-          }
-        } else {
-          resolve();
-        }
-      });
-    });
-
-    const file = (req as any).file;
-    if (!file) throw new BadRequestException('Archivo requerido');
-
-    return this.importGatewayService.processFile(file, name, userId);
   }
 
   @MessagePattern('ftp.removeFile')
